refactored comparison functions
